Card styling fix~
This fixed the card layouts on the index page - before, it was only displaying a centered single column of the recipe cards. This changed it to have a fixed set of 3 per row. One small fix is when the recipes are displayed on smaller screens, it needs to change to single recipes since it just squishes the elements onto each other.

Added a partial for the index cards as well.