Created dashboard
Starting dashboard template for a logged in user.
This includes the routes, controller, and mostly empty index and show erb files for the time being. My thoughts are to have the logged in "current user" land on the index page with the info on Figjam, then "show" can just be their own recipes