Feature spec for recipe edit action
This finishes updating all specs for the updated recipe controller/model to reflect the description assertion (I know this is a bigger PR) I definitely need to do separate branches in these situations in the future. BUT this adds assertions for correct rendering of the edit page and for an invalid scenario with error message.