Dash recipes fix
This fixes the weird styling issue on the user dashboard to have consistent styling with the recipe cards, and randomly selects from 3 recipes for the iteration for specificity. This also fixes the second part of the carousel that displayed the sample cards versus the actual favorited recipes. It now consistently displays 6 randomly generated favorited recipes for the user dash.