Recipe deletion
This update allows the current user (obviously the validations here will be needed later) to delete a recipe via the recipe's show page (with tests!). This currently is wired up to go back to the root path (home page) and has a generic message built into Turbo, which I'd like to change in a separate ticket to have a successful deletion message on redirect.